producer_service/requirements.txt
fastapi
pydantic
kafka-python

uvicorn
